// src/app/model/users.type.ts
type Info = {
  id: number;
  login: string;
  nome: string;
  email: string;
  password: string;
  morada: string;
  caixaPostal: string;
  pais: string;
};

export type Users = {
  id?: number;
  login: string;
  nome: string;
  email: string;
  password: string;
  morada: string;
  caixaPostal: string;
  pais: string;
};